feat: extract shard propagation pass
